requirements.txt
openai
python-dotenv
PyMuPDF
openapi-schema-validator
PyYAML
